RSA-In-Detail/requirements.txt
fastapi==0.111.*
uvicorn==0.30.*
jwcrypto==1.5.*
cryptography==41.*
PyJWT==2.8.*